vetka-mcp-full/requirements.txt
# VETKA MCP Full — Complete Requirements
# Full MCP server with all required modules

# ===== Core API / Server =====
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.6
sse-starlette>=1.6.0

# ===== HTTP & Config =====
httpx>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0
orjson>=3.9.0

# ===== MCP SDK =====
mcp>=1.0.0

# ===== Utilities =====
click>=8.1.0
tqdm>=4.66.0
tabulate>=0.9.0

# ===== Memory Stack =====
qdrant-client>=1.10.0
weaviate-client>=4.0.0
numpy>=2.0.0
scikit-learn>=1.4.0

# ===== Orchestration =====
langchain-core>=0.1.0
networkx>=3.0

# ===== Agent / Role Generation =====
jinja2>=3.1.0

# ===== Testing (optional) =====
pytest>=7.0.0
pytest-asyncio>=0.21.0
